Implemented CRUD operations in UserController.cs and updated WebGrease binding
Implemented the `Index`, `Details`, `Create`, `Edit`, and `Delete` methods in `UserController.cs` to handle CRUD operations for a `User` object. Updated the `bindingRedirect` for `WebGrease` in the `Web.config` file to redirect any version from `0.0.0.0` to `1.6.5135.21930` to version `1.6.5135.21930`. Also, reorganized the reference to `Microsoft.CodeDom.Providers.DotNetCompilerPlatform` in the `CRUD application 2.csproj` file for better organization. A redundant change to the closing `Project` tag in the same file was also made.